requirements.txt
requests
google-cloud-firestore==1.4.0
google-cloud-error-reporting==0.32.1
spotipy
numpy
scipy
sklearn